Extend Logical interpreter
