+++ moses/atomese/interpreter/Interpreter.h
@@ -27,6 +27,7 @@
 #include <opencog/atoms/base/Handle.h>
 #include <opencog/atoms/proto/LinkValue.h>
 #include <boost/iterator/zip_iterator.hpp>
+#include <moses/atomese/interpreter/logical_interpreter.h>
 
 /**
  * class Interpreter -- create an unfolded program from compressed program
@@ -64,41 +65,42 @@ private:
 	ProtoAtomPtr execute(Type t, ProtomSeq &params);
 };
 
-struct zip_adder :
-		public std::unary_function<const boost::tuple
-				<const ProtoAtomPtr&, const ProtoAtomPtr&>&, void>
-{
-	std::vector<ProtoAtomPtr> _result;
-
-	void operator()(const boost::tuple<const ProtoAtomPtr&, const ProtoAtomPtr&>& t)
-	{
-		if(HandleCast(t.get<0>())->get_type() == TRUE_LINK
-		   && HandleCast(t.get<1>())->get_type() == TRUE_LINK){
-			_result.push_back(ProtoAtomPtr(createLink(TRUE_LINK)));
-		}
-		else{
-			_result.push_back(ProtoAtomPtr(createLink(FALSE_LINK)));
-		}
-	}
-};
-
 LinkValuePtr logical_and(const LinkValuePtr& p1, const LinkValuePtr& p2){
 	std::vector<ProtoAtomPtr> p1_value = p1->value();
 	std::vector<ProtoAtomPtr> p2_value = p2->value();
 
-	zip_adder adder = std::for_each(
+	zip_and _and = std::for_each(
 			boost::make_zip_iterator(
 					boost::make_tuple(p1_value.begin(), p2_value.begin())
 			),
 			boost::make_zip_iterator(
 					boost::make_tuple(p1_value.end(), p2_value.end())
 			),
-			zip_adder()
+			zip_and()
 	);
 
-	std::vector<ProtoAtomPtr> _result = adder._result;
+	std::vector<ProtoAtomPtr> _result = _and._result;
 	return LinkValuePtr(new LinkValue(_result));
 }
+
+LinkValuePtr logical_or(const LinkValuePtr& p1, const LinkValuePtr& p2){
+	std::vector<ProtoAtomPtr> p1_value = p1->value();
+	std::vector<ProtoAtomPtr> p2_value = p2->value();
+
+	zip_or _or = std::for_each(
+			boost::make_zip_iterator(
+					boost::make_tuple(p1_value.begin(), p2_value.begin())
+			),
+			boost::make_zip_iterator(
+					boost::make_tuple(p1_value.end(), p2_value.end())
+			),
+			zip_or()
+	);
+
+	std::vector<ProtoAtomPtr> _result = _or._result;
+	return LinkValuePtr(new LinkValue(_result));
+}
+
 }
 }
